+++ src/main/java/com/n1netails/n1netails/slack/model/SlackElement.java
@@ -1,0 +1,7 @@
+package com.n1netails.n1netails.slack.model;
+
+import com.slack.api.model.block.element.BlockElement;
+
+public interface SlackElement {
+    BlockElement toBlockElement();
+}
